requirements.txt
# Core LangChain & LangGraph
langgraph>=0.0.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.1
anthropic>=0.18.0

# Vector Database
chromadb>=0.4.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
rich>=13.0.0